% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brnet.R
\name{adjtodist}
\alias{adjtodist}
\title{Convert an adjacency matrix to a distance matrix}
\usage{
adjtodist(x)
}
\arguments{
\item{x}{Adjacency matrix to be converted}
}
\description{
Convert an adjacency matrix to a distance matrix
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
