% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcbrnet.R
\name{brnet}
\alias{brnet}
\title{Generate a random branching network}
\usage{
brnet(
  n_patch = 50,
  p_branch = 0.5,
  mean_env_source = 0,
  sd_env_source = 1,
  rho = 1,
  sd_env_lon = 0.1,
  mean_disturb_source = 0.9,
  sd_disturb_source = 1,
  sd_disturb_lon = 0.1,
  randomize_patch = TRUE,
  n_patch_free = FALSE
)
}
\arguments{
\item{n_patch}{Number of patches in a network.}

\item{p_branch}{Branching probability
(success probability of a geometric distribution).}

\item{mean_env_source}{Mean value of environmental condition
at upstream origins of water.}

\item{sd_env_source}{SD of environmental condition
at upstream origins of water.}

\item{rho}{Strength of spatial autocorrelation in environmental condition.}

\item{sd_env_lon}{SD of longitudinal environmental noise.
The SD is defined in a logit scale with a normal distribution.}

\item{mean_disturb_source}{Mean disturbance strength
at upstream origins of water.
The value represents the proportional mortality (0 - 1.0).}

\item{sd_disturb_source}{SD of disturbance strength
at upstream origins of water.
The SD is defined in a logit scale with a normal distribution.}

\item{sd_disturb_lon}{SD of longitudinal noise of disturbance strength.
The SD is defined in a logit scale with a normal distribution.}

\item{randomize_patch}{Logical
indicating whether patches are randomized or not.
If \code{FALSE}, the function may generate a topologically-biased network.
Default \code{TRUE}.}

\item{n_patch_free}{Logical indicating
whether imposing a constraint on \code{n_patch} or not.
If \code{TRUE}, the number of patches is a random variable
following a negative binomial distribution.}
}
\description{
Generate a random branching network
}
\section{Reference}{
 see \href{https://aterui.github.io/mcbrnet/}{package webpage} for instruction
}

\examples{
\dontrun{
# not run
brnet(n_patch = 10, p_branch = 0.5)
}

}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
