% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_b.R
\name{calc_b}
\alias{calc_b}
\title{Calculate density dependent growth parameter}
\usage{
calc_b(r_max, k)
}
\arguments{
\item{r_max}{maximum per capita recruitment rate for basal species B}

\item{k}{carrying capacity}
}
\value{
parameter b
}
\description{
Internal function used by \code{k_function_internal} to calculate parameter \code{b}, which determines asymptotic level of basal species, B, recruitment
}
\details{
b = (r_max - 1) / k
}
\examples{
calc_b(r_max = 2, k = 100)

}
