% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disturb_internal.R
\name{disturb_internal}
\alias{disturb_internal}
\title{Determine if a disturbance occurred and reduce population abundances in a time step.}
\usage{
disturb_internal(
  N,
  disturb_type,
  disturb_p,
  disturb_value,
  river_network_structure
)
}
\arguments{
\item{N}{Abundance matrix with \code{nrow = n_sp = 3} and \code{ncol = n_patch}.}

\item{disturb_type}{Character vector of one of \verb{c("regional", or NULL)}. \code{"regional"} applies a disturbance to all patches within the network.
If \code{disturb_type = NULL}, no disturbances will occur in the simulation.
Future releases are intended to include an additional disturbance type for branching river networks: \code{"point-source"}. Which is intended to apply a disturbance randomly to individual patches, and the magnitude of disturbance will decay downstream. This is not currently implemented.}

\item{disturb_p}{probability of disturbance occurring. If \code{disturb_type = "regional"}, disturbance for the entire meta-community is determined in each time step via \code{rbinom(n = 1, size = 1, prob = disturb_p)}.}

\item{disturb_value}{Vector of numbers from 0 to 1 describing the magnitude of disturbance \code{i}.
This is optimized to work with branching river networks generated from the the \code{mcbrnet} package by taking the output of \code{df_patch$disturbance} from \code{brnet()} as an input. See \code{?brnet} for more details.
For 2D habitats, this number is sampled randomly by converting the \code{disturb_mag_mean} and \code{disturb_mag_sd} arguments using the inverse logit function. Currently, the value for each patch is randomly sampled, i.e., there is no spatial correlation. This has not been optimized for 2D habitats}

\item{river_network_structure}{Logical indicating if habitat architecture is branching (\code{TRUE}) or a 2D square (\code{FALSE})}
}
\value{
List with two elements. \code{N} = abundance matrix after accounting for disturbances. Values don't necessarily have to be integers. Numeric-double values will be converted to integer with \code{rpois()} in the simulation function. \code{patch_extinction} is an integer vector of \code{length = n_patch} indicating if a disturbance did happen (1) or did not happen (0).
}
\description{
Determine if a disturbance occurred and reduce population abundances in a time step.
}
\examples{
disturb_internal(N, disturb_type, disturb_p, disturb_value, river_network_structure)

}
