% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_to_k.R
\name{env_to_k}
\alias{env_to_k}
\title{Calculate K based on environmental value}
\usage{
env_to_k(env, k_base)
}
\arguments{
\item{env}{Vector of environmental values. Inherited from \code{igp_sim()} function.}

\item{k_base}{Mean carrying capacity K.}
}
\value{
Vector of carrying capacities
}
\description{
function used internally by \code{k_function_internal} when \code{igp_sim(k_function = "environment")}. Inherits all arguments from \code{igp_sim()}.
}
\details{
This function adds variation to the carrying capacity based on the environment value. When the variation in the \code{env} variable is high, the carrying capacity varies between 0.5 and 1.5 * \code{k_base}, with most of the density being concentrated at the extremes. When variation in \code{env} is very low, the observed k is approximately = \code{k_base}.

This function is designed to take the \code{df_patch$environment} output from the \code{mcbrnet::brnet()} as the \code{env} argument. See the \code{brnet()} function to control the variation of \code{environment} in head waters as well as longitudinal variation of environmental values between patches.
}
\examples{
env_to_k(env = rnorm(10), k_base = 100)

}
