% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm.R
\name{extra_prey}
\alias{extra_prey}
\title{Extra prey function}
\usage{
extra_prey(A, j, i0, tp, theta, kappa, cannibal = FALSE)
}
\arguments{
\item{A}{Intreaction matrix}

\item{j}{Consumer's index}

\item{i0}{First prey's index}

\item{tp}{Initial trophic positions}

\item{theta}{Numeric. Scale parameter of an exponential distribution. Smaller values indicate greater trophic specialization.}

\item{kappa}{Number of extra prey items}

\item{cannibal}{Logical. If \code{TRUE}, cannibalism allowed}
}
\description{
Extra prey function
}
