% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sglv.R
\name{findr}
\alias{findr}
\title{Find suitable intrinsic growth rates}
\usage{
findr(alpha, k0, lambda0 = 0.001, interval = 0.01)
}
\arguments{
\item{alpha}{Numeric. n_species x n_species interaction matrix}

\item{k0}{Numeric. Total carrying capacity for basal species combined}

\item{lambda0}{Numeric. Initial lambda value for the expential decay of equilibrium densities with trophic position}

\item{interval}{Numeric. Increment of lambda value.}
}
\description{
Find suitable intrinsic growth rates
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
