% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcbrnet.R
\name{foodchain}
\alias{foodchain}
\title{Spatiotemporal average of maximum trophic positions using \code{\link{sglv}} output}
\usage{
foodchain(
  n,
  n_species,
  n_patch,
  alpha,
  start = ceiling(max(n[, 1]) - max(n[, 1]) * 0.5) + 1,
  end = floor(max(n[, 1])),
  full = FALSE
)
}
\arguments{
\item{n}{Numeric. Matrix output from \code{\link{sglv}}}

\item{n_species}{Integer. Number of species}

\item{n_patch}{Integer. Number of habitat patches}

\item{alpha}{Numeric. Interaction matrix}

\item{start}{Integer. Initial time step to be included in the calculation. \code{floor} applied internally}

\item{end}{Integer. Last time step to be included in the calculation. \code{floor} applied internally}

\item{full}{Logical. If \code{TRUE}, a full spatiotemporal matrix of food chain length returned as attributes.}
}
\description{
Spatiotemporal average of maximum trophic positions using \code{\link{sglv}} output
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
