% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm.R
\name{foodweb}
\alias{foodweb}
\title{Generate a food web based on the preferential prey model}
\usage{
foodweb(
  n_species,
  n_basal,
  l,
  theta,
  cannibal = FALSE,
  attack = list(min = 0, max = 1),
  convert = list(min = 0, max = 1),
  mortal = list(min = 0, max = 1)
)
}
\arguments{
\item{n_species}{Integer. Number of species}

\item{n_basal}{Integer. Number of basal species}

\item{l}{Interger. Expected number of links in the upper triangle}

\item{theta}{Numeric. Scale parameter of an exponential distribution. Smaller values indicate greater trophic specialization.}

\item{cannibal}{Logical. If \code{TRUE}, cannibalism allowed}

\item{attack}{List for attack rates. Specify minimum and maximum values for a uniform distribution.}

\item{convert}{List for conversion efficiency. Specify minimum and maximum values for a uniform distribution.}

\item{mortal}{List for mortality (or intraspecific competition). Specify minimum and maximum values for a uniform distribution.}
}
\description{
Generate a food web based on the preferential prey model
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
