% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fun_adj}
\alias{fun_adj}
\title{Return adjacency in a linear habitat}
\usage{
fun_adj(n, start_id = 1)
}
\arguments{
\item{n}{Number of nodes}

\item{start_id}{Start node ID}
}
\description{
Return adjacency in a linear habitat
}
