% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fun_dispersal}
\alias{fun_dispersal}
\title{Dispersal}
\usage{
fun_dispersal(x, v_p_dispersal, m_dispersal)
}
\arguments{
\item{x}{Population size matrix}

\item{v_p_dispersal}{Vector of dispersal probability}

\item{m_dispersal}{Dispersal matrix}
}
\description{
Dispersal
}
