% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get_fcl.R
\name{fun_get_fcl}
\alias{fun_get_fcl}
\title{Internal function: get fcl}
\usage{
fun_get_fcl(x, delta)
}
\arguments{
\item{x}{Community matrix}

\item{delta}{Preference to basal over ig-prey}
}
\description{
Internal function: get fcl
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
