% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fun_get_n_branch}
\alias{fun_get_n_branch}
\title{Get n_branch}
\usage{
fun_get_n_branch(n_patch, p_branch)
}
\arguments{
\item{n_patch}{Number of patches in a network.}

\item{p_branch}{Branching probability (success probability of a geometric distribution).}
}
\description{
Get n_branch
}
