% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_patch_attr.R
\name{fun_patch_attr}
\alias{fun_patch_attr}
\title{Internal function: weighted-mean patch attributes}
\usage{
fun_patch_attr(
  x,
  n_branch,
  mean_source,
  sd_source,
  sd_lon,
  m_distance,
  rho = 1,
  v_wa,
  logit = FALSE
)
}
\arguments{
\item{x}{adjacency matrix to be converted}

\item{n_branch}{number of branch}

\item{mean_source}{mean value of source attribute}

\item{sd_source}{SD of source attribute}

\item{sd_lon}{SD of longitudinal change}

\item{m_distance}{distance matrix}

\item{rho}{longitudinal autocorrelation}

\item{v_wa}{vector of watershed area}

\item{logit}{logit transformation of mean source attribute}
}
\description{
Internal function: weighted-mean patch attributes
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
