% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fun_to_m}
\alias{fun_to_m}
\title{To matrix}
\usage{
fun_to_m(x, n_species, n_patch, param_attr)
}
\arguments{
\item{x}{Scalar or vector}

\item{n_species}{Number of species}

\item{n_patch}{Number of patches}

\item{param_attr}{Indicate species or patch attribute}
}
\description{
To matrix
}
