% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_to_v.R
\name{fun_to_v}
\alias{fun_to_v}
\title{Internal function: to vector}
\usage{
fun_to_v(x, n)
}
\arguments{
\item{x}{scalar or vector}

\item{n}{replication (n_species or n_patch)}
}
\description{
Internal function: to vector
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
