% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_wa.R
\name{fun_wa}
\alias{fun_wa}
\title{Internal function: watershed area}
\usage{
fun_wa(x, n_patch)
}
\arguments{
\item{x}{Adjacency matrix to be converted}
}
\description{
Internal function: watershed area
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
