% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fcl_state.R
\name{get_fcl_state}
\alias{get_fcl_state}
\title{Calculates FCL "state" variable}
\usage{
get_fcl_state(N)
}
\arguments{
\item{N}{Matrix of population abundances, where \code{nrow = 3}, and \code{ncol = n_patch}.}
}
\value{
numeric vector of length = \code{n_patch}. FCL state 0 = no species, 1 = B only, 2 = B + C, 2.5 = B + P, 3 = B + C + P
}
\description{
Internal function used by \code{igp_sim()} to describe community composition in each patch for each time step
}
\details{
This function calculates the food chain length state of each patch at each time step in the IGP simulation. This is used internally.
}
\examples{
# simulate abundance matrix
# nrow = n_sp = 3
# ncol = n_patch = 10
N <- matrix(rpois(3 * 10, lambda = 1), nrow = 3, ncol = 10)

# make sure abundance of C and P = 0 when abundance of B = 0
N[,N[1,] == 0] <- 0

# calculate food chain length state
get_fcl(N)

}
