% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_ggbrnet.R
\name{ggbrnet}
\alias{ggbrnet}
\title{Visualize a branching network}
\usage{
ggbrnet(
  adjacency_matrix,
  patch_attr,
  patch_color = "env",
  patch_label = "none",
  patch_size = 3,
  ...
)
}
\arguments{
\item{adjacency_matrix}{Adjacency matrix.}

\item{patch_attr}{Data frame for patch attributes.}

\item{patch_color}{Type of patch (vertex) label (either \code{"env"}, \code{"disturbance"} or any color code). Default \code{"env"}.}

\item{patch_label}{Type of patch (vertex) label (either \code{"patch", "branch", "n_upstream"}). \code{"patch"} shows patch ID, \code{"branch"} branch ID, and \code{"n_upstream"} the number of upstream contributing patches. If \code{"none"}, no label will be shown on patches in the plot. Default \code{"none"}.}

\item{patch_size}{Patch (vertex) size in the plot.}

\item{...}{Arguments passed to \code{ggraph::geom_node_label()}.}
}
\description{
Visualize a branching network
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
