% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_function_internal.R
\name{k_function_internal}
\alias{k_function_internal}
\title{Internal function used to calculate carrying capacity within \code{igp_sim()}}
\usage{
k_function_internal(
  k_function,
  k_base,
  r_max,
  n_upstream,
  n_patch,
  k_c,
  k_min_exponent,
  k_max_exponent,
  river_network_structure,
  environment_value
)
}
\arguments{
\item{k_function}{Character string, one of c("patches upstream", "environment", or NULL).}

\item{k_base}{Numeric input scaling the carrying capacity. See details for more information.}

\item{r_max}{Maximum reproductive rate of basal species, B}

\item{n_upstream}{Vector of length = n_patch describing the number (n+1) of upstream patches. Terminal upstream patches have a value of 1.}

\item{n_patch}{The number of total patches}

\item{k_c}{constant (default = 10) used in \code{k_n_upstream()} when \code{k_function =  "patches-upstream"}}

\item{k_min_exponent}{min exponent (default = 1.1) used in \code{k_n_upstream()} when \code{k_function =  "patches upstream"}}

\item{k_max_exponent}{min exponent (default = 1.35) used in \code{k_n_upstream()} when \code{k_function =  "patches upstream"}}

\item{river_network_structure}{Logical indicating if this is a branching river network \code{TRUE} or not \code{FALSE}.}

\item{environment_value}{Vector of environmental values. Designed to take \code{brnet()} function output.}
}
\value{
\code{k} vector of length = \code{n_patch} which describes carrying capacity for each patch
\code{b} parameter describing density dependent growth of basal species B.
}
\description{
Function used internally by \code{igp_sim()}. Inherits all arguments from \code{igp_sim()} function call, or others calculated internally.
}
\examples{
k_function_internal(k_function, k_base, r_max, n_upstream, n_patch, k_c, k_min_exponent, k_max_exponent, river_network_structure, environment_value)

}
