% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_n_upstream.R
\name{k_n_upstream}
\alias{k_n_upstream}
\title{Carrying Capacity for basal species B, based on number of upstream nodes}
\usage{
k_n_upstream(
  k_base = 150,
  k_c = 10,
  k_min_exponent = 1.1,
  k_max_exponent = 1.35,
  r_max,
  n_upstream,
  n_patch
)
}
\arguments{
\item{k_base}{integer}

\item{k_c}{numeric}

\item{k_min_exponent}{minimum value for exponent describing productivity ~ watershed area relationship. default is 1.10 (Koening et al. 2019) numeric}

\item{k_max_exponent}{minimum value for exponent describing productivity ~ watershed area relationship. default is 1.35 (Koening et al. 2019) numeric}

\item{r_max}{maximum reproductive rate, numeric}

\item{n_upstream}{number of patches upstream + 1. i.e., the most upstream terminal patch has \code{n_upstream = 1}, integer}

\item{n_patch}{number of patches to calculate, should be same length as n_upstreamm, integer}
}
\value{
list with 2 vector elements, each the length of \code{n_patch}. \code{k} = carrying capacity for that patch. \code{b} = parameter controlling realized strength of population growth at a given population size. b is used internally in \code{igp_sim()}
}
\description{
internal function used by \code{k_function_internal} when \code{igp_sim(k_function = "patches-upstream")}. Inherits all arguments from \code{igp_sim()}.
}
\details{
The minimum carrying capacity is equal to \code{k_base} + \code{k_c}. The upstream terminal node The exponent for each patch is sampled from a uniform distribution from \code{k_min_exponent} to \code{k_max_exponent}.
k = k_base + k_c * n_upstream^k_exp
K is automatically rounded to the nearest integer.
b = (r_max - 1) /k
}
\examples{
k_n_upstream(k_base = 150, k_c = 10, k_min_exponent = 1.10,
 k_max_exponent = 1.35, r_max = 4,
  n_upstream = c(5, 10), n_patch = 2)

}
\references{
Koenig, L.E., Helton, A.M., Savoy, P., Bertuzzo, E., Heffernan, J.B., Hall, R.O., Jr. and Bernhardt, E.S. (2019), Emergent productivity regimes of river networks. Limnol Oceanogr, 4: 173-181. https://doi.org/10.1002/lol2.10115
}
