% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcbrnet.R
\name{max_tp}
\alias{max_tp}
\title{Maximum trophic position}
\usage{
max_tp(n_species, n, alpha)
}
\arguments{
\item{n_species}{Integer. Number of species}

\item{n}{Numeric. Vector of species density}

\item{alpha}{Numeric. Interaction matrix}
}
\description{
Maximum trophic position
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
