% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcbrnet-package.R
\docType{package}
\name{mcbrnet-package}
\alias{mcbrnet}
\alias{mcbrnet-package}
\title{mcbrnet: Ecological simulation in spatial networks}
\description{
Ecological simulation in spatial networks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://aterui.github.io/mcbrnet/}
}

}
\keyword{internal}
