% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcsim.R
\name{mcsim}
\alias{mcsim}
\title{Simulate metacommunity dynamics}
\usage{
mcsim(
  n_species = 5,
  n_patch = 5,
  n_warmup = 200,
  n_burnin = 200,
  n_timestep = 1000,
  propagule_interval = NULL,
  propagule_seed = 0.5,
  carrying_capacity = 100,
  xi = 1,
  interaction_type = "constant",
  alpha = 0,
  min_alpha = NULL,
  max_alpha = NULL,
  r0 = 4,
  niche_optim = NULL,
  min_optim = -1,
  max_optim = 1,
  sd_niche_width = NULL,
  min_niche_width = 0.1,
  max_niche_width = 1,
  niche_cost = 1,
  xy_coord = NULL,
  distance_matrix = NULL,
  dispersal_matrix = NULL,
  p_disturb = 0,
  i_disturb = 0,
  landscape_size = 10,
  mean_env = 0,
  sd_env = 0.1,
  spatial_env_cor = FALSE,
  phi = 1,
  p_dispersal = 0.1,
  theta = 1,
  zeta = 0,
  impact = 0,
  plot = FALSE
)
}
\arguments{
\item{n_species}{Number of species in a metacommunity.}

\item{n_patch}{Number of patches in a metacommunity.}

\item{n_warmup}{Number of time-steps for warm-up. Default \code{200}.}

\item{n_burnin}{Number of time-steps for burn-in. Default \code{200}.}

\item{n_timestep}{Number of time-steps to be saved. Default \code{1000}.}

\item{propagule_interval}{Time interval for propagule introduction during warm-up. If \code{NULL}, a value of \code{ceiling(n_warmup / 10)} will be used.}

\item{propagule_seed}{Propagule mean density (intensity parameter in a Poisson distribution). Default \code{0.5}.}

\item{carrying_capacity}{Carrying capacities of individual patches. Length must be one or equal to \code{n_patch}. Default \code{100}.}

\item{xi}{Hassell exponent. Undercompensation (xi < 1), no compensation (xi = 1; reduced to Beverton-Holt), and overcompensation (xi > 1).}

\item{interaction_type}{\code{"constant"} or \code{"random"}. \code{"constant"} assumes the unique interaction strength of \code{alpha} for all pairs of species. \code{"random"} draws random numbers from a uniform distribution with \code{min_alpha} and \code{max_alpha}.}

\item{alpha}{Species interaction strength. Enabled if \code{interaction_type = "constant"}. Default \code{0}.}

\item{min_alpha}{Minimum value of a uniform distribution that generates species interaction strength. Enabled if \code{interaction_type = "random"}. Default \code{NULL}.}

\item{max_alpha}{Maximum value of a uniform distribution that generates species interaction strength. Enabled if \code{interaction_type = "random"}. Default \code{NULL}.}

\item{r0}{Maximum reproductive number of the Beverton-Holt model. Length must be one or equal to \code{n_species}.}

\item{niche_optim}{Niche optimum of species (environmental value that maximizes the reproductive number). Length must be one or equal to \code{n_species}. Default \code{NULL}.}

\item{min_optim}{Minimum value of a uniform distribution that generates optimal environmental values of simulated species. Values are randomly assigned to species. Enabled if \code{niche_optim = NULL}.}

\item{max_optim}{Maximum value of a uniform distribution that generates optimal environmental values of simulated species. Values are randomly assigned to species. Enabled if \code{niche_optim = NULL}.}

\item{sd_niche_width}{Niche width of species. Higher values indicate greater niche width. Length must be one or equal to \code{n_species}.}

\item{min_niche_width}{Minimum value of a uniform distribution that generates niche width values of simulated species. Values are randomly assigned to species. Enabled if \code{sd_niche_width = NULL}.}

\item{max_niche_width}{Maximum value of a uniform distribution that generates niche width values of simulated species. Values are randomly assigned to species. Enabled if \code{sd_niche_width = NULL}.}

\item{niche_cost}{Determine the cost of wide niche (smaller values imply greater costs of wider niche). Default \code{1}.}

\item{xy_coord}{Site coordinates. Must be provided as a data frame in which each row corresponds to an individual site with x and y coordinates (columns). Defualt \code{NULL}.}

\item{distance_matrix}{Distance matrix indicating distance between habitat patches. If provided, the distance matrix will be used to generate dispersal matrix and to calculate distance decay of environmental correlations. Default \code{NULL}.}

\item{dispersal_matrix}{Dispersal matrix to be used to simulate dispersal process. Override distance_matrix. Default \code{NULL}.}

\item{p_disturb}{Disturbance probability.}

\item{i_disturb}{Disturbance intensity expressed as proportional mortality (0 to 1). Length must be one or equal to \code{n_patch}.}

\item{landscape_size}{Length of a landscape on a side. Enabled if \code{dispersal_matrix = NULL}.}

\item{mean_env}{Mean environmental values of patches. Length must be one or equal to \code{n_patch}.}

\item{sd_env}{Standard deviation of temporal environmental variation at each patch.}

\item{spatial_env_cor}{If \code{TRUE}, spatial autocorrelation in temporal environmental fluctuation is considered. Default \code{FALSE}.}

\item{phi}{Parameter describing the distance decay of spatial autocorrelation in temporal environmental fluctuation. Enabled if \code{spatial_env_cor = TRUE}.}

\item{p_dispersal}{Dispersal probability. Length must be one or equal to \code{n_species}.}

\item{theta}{Rate parameter of exponential dispersal kernel.}

\item{zeta}{Species sensitivity to environmental pollutants.}

\item{impact}{Concentration of environmental pollutants.}

\item{plot}{If \code{TRUE}, five sample patches and species of \code{df_dynamics} are plotted.}
}
\description{
Simulate metacommunity dynamics
}
\section{Reference}{
 see \href{https://aterui.github.io/mcbrnet/index.html}{package webpage} for instruction
}

\examples{
\dontrun{
# not run
mcsim(n_patch = 5, n_species = 5)
}

}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
