% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm.R
\name{ppm}
\alias{ppm}
\title{Preferential prey model}
\usage{
ppm(n_species, n_basal, l, theta, cannibal = FALSE)
}
\arguments{
\item{n_species}{Integer. Number of species}

\item{n_basal}{Integer. Number of basal species}

\item{l}{Interger. Expected number of links in the upper triangle}

\item{theta}{Numeric. Scale parameter of an exponential distribution. Smaller values indicate greater trophic specialization.}

\item{cannibal}{Logical. If \code{TRUE}, cannibalism allowed}
}
\description{
Preferential prey model
}
