% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prey_preference.R
\name{prey_preference}
\alias{prey_preference}
\title{Calculate preference between two prey resources.}
\usage{
prey_preference(e1, e2, N1, N2)
}
\arguments{
\item{e1}{Conversion efficiency for prey resource 1. Most often this \code{ebp}, or the efficiency of turning B biomass into new P biomass.}

\item{e2}{Conversion efficiency for prey resource 2. Most often this \code{ecp}, or the efficiency of turning C biomass into new P biomass.}

\item{N1}{The population abundance of prey resource 1. Usually this is prey B}

\item{N2}{The population abundance of prey resource 2. Usually this is prey C}
}
\value{
value between 0 and 1 describing search preference for resource B by predator P.
}
\description{
Internal function used by \code{pop_sim()} when \code{igp_sim(P_pref = NULL)}
}
\details{
The preference of prey resource 1 over prey resource 2. Values < 0.5 indicate the resource 2 is preferred, value > 0.5 indicate resource 1 is preferred.

Preference is calculated as \code{e1*N1 / (e1*N1 + e2*N2)}
}
\examples{
prey_preference(e1 = 2, e2 = 1, N1 = 100, N2 = 50)

}
