% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcbrnet.R
\name{ptsource}
\alias{ptsource}
\title{Create point source disturbance}
\usage{
ptsource(x, n_source, p, q, pattern = "random")
}
\arguments{
\item{x}{\code{brnet()} object}

\item{n_source}{Numeric. Number of point sources}

\item{p}{Numeric. Reduction factor for downstream propagation}

\item{q}{Numeric. Reduction factor for upstream propagation}

\item{pattern}{Character.
Spatial pattern of point sources.
Either "random", "cluster", "upstream", or "downstream"}
}
\description{
Create point source disturbance
}
\section{Reference}{

see \href{https://aterui.github.io/mcbrnet/}{package webpage} for details
}

\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
