% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_k_sp.R
\name{r_k_sp}
\alias{r_k_sp}
\title{Maximum reproductive rate and carrying capacity for all three species}
\usage{
r_k_sp(
  k_b = 150,
  r_b = 2.5,
  ebc = 2,
  ebp = 2,
  ecp = 2,
  alphabc = 4,
  alphap = 4,
  betabc = 20,
  betap = 20,
  P_pref = NULL
)
}
\arguments{
\item{k_b}{Carrying capacity of basal species B}

\item{r_b}{Maximum reproductive rate of basal species, B}

\item{ebc}{conversion efficiency of turning B biomass into new C biomass}

\item{ebp}{conversion efficiency of turning B biomass into new P biomass}

\item{ecp}{conversion efficiency of turning C biomass into new p biomass}

\item{alphabc}{parameter controlling predation and reproductive rates of consumer species, C}

\item{alphap}{parameter controlling predation and reproductive rates of consumer species, P}

\item{betabc}{parameter controlling predation and reproductive rates of consumer species, C}

\item{betap}{parameter controlling predation and reproductive rates of consumer species, P}

\item{P_pref}{The preference that species P has for B over C. Default is \code{NULL}, and will be calculated based on other parameters (see below). This value can be fixed by the user i.e. \code{P_pref = 0.25} indicates that the predator spends ~25\% of it's time searching for B, and \verb{1 - P_pref = 75\%}  of its time searching for C.}
}
\value{
a list with 3 named elements: \code{max_r} which has the maximum reproductive rates for B (\code{r_b}) and C (\code{r_c}), and the partitioned (\code{r_bp}, and \code{r_cp}) and total rates for P (\code{r_p_total}); \code{k} which has the carrying capacity for B (\code{k_b}) and C (\code{k_c}), and the partitioned and total rates for P; \code{input} which has the input values used for calculations.
}
\description{
This is a helper function to estimate reproductive rates and carrying capacity of species C and P.
}
\details{
This function calculates the maximum reproductive rate and carrying capacity for each species in the IGP simulation communities. This can be used to estimate what effect different parameters will have. Carrying capacities are calculated assuming prey resources are at their respective carrying capacities, hence the return values estimate upper ceiling of population abundances for a given set of parameters. This function does also not account for indirect effects, i.e. C predation upon B lowering resource abundance for P.

Carrying capacity and maximum reproductive rate of the basal species is set by the user.

Maximum reproductive rates for C are calculated as \code{r_c = ebc * alphac}. Carrying capacity is calculated as \code{k_c = k_b*(r_c - 1) / betac}

Maximum reproductive rates for the predator are calculated separately for both of its prey, and depends on its preference of prey resource B over C. If \code{P_pref = NULL}, the default, the preference is calculated according to the \code{prey_preference()} function.
}
\examples{
r_k_sp(k_b = 500, r_b = 2.5, ebc = 2,ebp = 2, ecp = 2, alphac = 4, alphap = 4, betac = 20, betap = 20, P_pref = NULL)

}
\seealso{
\code{prey_preference}
}
