% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_resample.R
\name{resample}
\alias{resample}
\title{Internal function: resample}
\usage{
resample(x, ...)
}
\arguments{
\item{x}{Vector}

\item{...}{Additional arguments passed to \code{sample}}
}
\description{
Internal function: resample
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
