% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcbrnet.R
\name{sglv}
\alias{sglv}
\title{Spatial generalized Lotka-Volterra model}
\usage{
sglv(
  n_species,
  n_patch,
  n_timestep = 100,
  interval = 0.01,
  r,
  alpha,
  dispersal = list(adj = matrix(0, nrow = n_patch, ncol = n_patch), phi = 0, m = 0),
  n0 = list(min = 0, max = 1),
  threshold = 0,
  cpp = TRUE,
  ...
)
}
\arguments{
\item{n_species}{Integer. Number of species}

\item{n_patch}{Integer. Number of habitat patches}

\item{n_timestep}{Integer. Number of time step for a simulation run}

\item{interval}{Numeric. Time interval used for a numerical ODE solver}

\item{r}{Numeric. Intrinsic growth rates of modeled species}

\item{alpha}{n_species x n_species interaction matrix}

\item{dispersal}{List. This list must contain the following parameters as named elements: \code{adj} n_patch x n_patch adjacency matrix; \code{phi} dispersal rate; \code{m} dispersal mortality rate.}

\item{n0}{List. Initial densities for modeled species that are randomly generated by \code{runif(n_species, min = min, max = max)}. This list must contain \code{min} and \code{max} as named elements.}

\item{threshold}{Numeric. Extinction threshold. Species will be removed from a simulation if species density goes below this value}

\item{cpp}{Logical. If \code{TRUE}, Rcpp function is used for the numerical solver}

\item{...}{Additional arguments passed to \code{\link{ode}}}
}
\description{
Spatial generalized Lotka-Volterra model
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
