% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_N.R
\name{sim_N}
\alias{sim_N}
\title{Simulate abundance matrix, N}
\usage{
sim_N(n_patch = 5, fcl_5_states = FALSE)
}
\arguments{
\item{n_patch}{Integer, number of patches (columns in N)}

\item{fcl_5_states}{Logical. If \code{TRUE}, \code{n_patch} must be >= 5, forces the first 5 columns of \code{N} to represent all possible community compositions in order: No species, B only, B+P, B+C, B+P+C.}
}
\value{
\code{N} abundance matrix with \code{nrow = 3} and \code{ncol = n_patch}.
}
\description{
This is a helper function to simulate an abundance matrix N. This is not used in the main package, and is included to trobleshoot problems and for illustrative purposes.
}
\examples{

# simulate N with all 5 community composition possibilities.
sim_N(n_patch = 5, fcl_5_states = TRUE)

}
