% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_brnet.R
\name{brnet}
\alias{brnet}
\title{Generate a random branching network}
\usage{
brnet(
  n_patch = 100,
  p_branch = 0.5,
  min_env = -1,
  max_env = 1,
  rho = 1,
  sd_env = 0.1,
  randomize_patch = TRUE,
  plot = TRUE
)
}
\arguments{
\item{n_patch}{number of patches in a network.}

\item{p_branch}{branching probability (success probability of a geometric distribution).}

\item{min_env}{minimum value of environmental condition in source streams (minimum of a uniform distribution).}

\item{max_env}{maximum value of environmental condition in source streams (maximum of a uniform distribution).}

\item{rho}{strength of spatial autocorrelation in environmental condition. The environmental condition at patch i \eqn{x}\out{<sub>i</sub>} is determined as \eqn{x}\out{<sub>i</sub>}\eqn{ = \rho}x\out{<sub>i-1</sub>}\eqn{ + \epsilon}\out{<sub>i</sub>}, where \eqn{\epsilon}\out{<sub>i</sub>} is the random variable drawn from a normal distribution with mean 0 and SD \eqn{\sigma}\out{<sub>env</sub>}.}

\item{sd_env}{SD of environmental noise (\eqn{\sigma}\out{<sub>env</sub>}).}

\item{randomize_patch}{logical indicating whether randomize patches or not. If \code{FALSE}, the function may generate a biased network with ordered patches. Default \code{TRUE}.}

\item{plot}{logical. If \code{FALSE}, a plot of the generated network will not be shown. Default \code{TRUE}.}
}
\value{
\code{adjacency_matrix} adjacency matrix for the generated network.

\code{distance_matrix} distance matrix for the generated network.

\code{environment} environmental values for patches.

\code{watershed_area} number of patches upstream. Patch 1 is set to be the root patch (i.e., outlet).

\code{branch_id} branch id for each patch.
}
\description{
Generate a random branching network
}
\examples{
brnet(n_patch = 100, p_branch = 0.5)

}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
