% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_brnet.R
\name{brnet}
\alias{brnet}
\title{Generate a random branching network}
\usage{
brnet(
  n_patch = 50,
  p_branch = 0.5,
  mean_env_source = 0,
  sd_env_source = 1,
  rho = 1,
  sd_env_lon = 0.1,
  mean_disturb_source = 0.9,
  sd_disturb_source = 1,
  sd_disturb_lon = 0.1,
  randomize_patch = TRUE,
  plot = FALSE,
  patch_color = "env",
  patch_label = "none",
  patch_size = 3,
  n_patch_free = FALSE,
  ...
)
}
\arguments{
\item{n_patch}{Number of patches in a network.}

\item{p_branch}{Branching probability (success probability of a geometric distribution).}

\item{mean_env_source}{Mean value of environmental condition at upstream terminals.}

\item{sd_env_source}{SD of environmental condition at upstream terminals.}

\item{rho}{Strength of spatial autocorrelation in environmental condition.}

\item{sd_env_lon}{SD of longitudinal environmental noise.}

\item{mean_disturb_source}{Mean disturbance strength at headwaters. The value is assumed to represent the proportional mortality (0 - 1.0) at the patch level.}

\item{sd_disturb_source}{SD of disturbance strength at headwaters. The SD is defined in a logit scale with a normal distribution.}

\item{sd_disturb_lon}{SD of longitudinal noise of disturbance strength. The SD is defined in a logit scale with a normal distribution.}

\item{randomize_patch}{Whether randomize patches or not. If \code{FALSE}, the function may generate a biased network with ordered patches. Default \code{TRUE}.}

\item{plot}{Whether a plot should be shown or not. If \code{FALSE}, a plot of the generated network will not be shown. Default \code{TRUE}.}

\item{patch_color}{Type of patch (vertex) label (either \code{"env"}, \code{"disturbance"} or any color code). Default \code{"env"}.}

\item{patch_label}{Type of patch (vertex) label (either \code{"patch", "branch", "n_upstream"}). \code{"patch"} shows patch ID, \code{"branch"} branch ID, and \code{"n_upstream"} the number of upstream contributing patches. If \code{"none"}, no label will be shown on patches in the plot. Default \code{"none"}.}

\item{patch_size}{Patch (vertex) size in the plot.}

\item{n_patch_free}{Whether imposing a constraint on \code{n_patch}. If \code{TRUE}, the number of patches a random variable following a negative binomial distribution.}

\item{...}{Arguments passed to \code{ggraph::geom_node_label()}.}
}
\description{
Generate a random branching network
}
\section{Reference}{
 see \href{https://aterui.github.io/mcbrnet/}{package webpage} for instruction
}

\examples{
\dontrun{
# not run
brnet(n_patch = 10, p_branch = 0.5)
}

}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
