% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_brnet.R
\name{brnet}
\alias{brnet}
\title{Generate a random branching network}
\usage{
brnet(
  n_patch,
  p_branch,
  mean_env_source = 0,
  sd_env_source = 1,
  rho = 1,
  sd_env_lon = 0.1,
  randomize_patch = TRUE,
  plot = TRUE,
  patch_label = NULL,
  patch_size = 6,
  patch_scaling = TRUE,
  scale_factor = 8
)
}
\arguments{
\item{n_patch}{numeric value indicating the number of patches in a network.}

\item{p_branch}{numeric value indicating the branching probability (success probability of a geometric distribution).}

\item{mean_env_source}{numeric value indicating the mean value of environmental condition at upstream terminals.}

\item{sd_env_source}{numeric value indicating the SD of environmental condition at upstream terminals.}

\item{rho}{numeric value indicating the strength of spatial autocorrelation in environmental condition. The environmental condition at patch x \eqn{z}\out{<sub>x</sub>} is determined as \eqn{z}\out{<sub>x</sub>}\eqn{ = \rho}z\out{<sub>x-1</sub>}\eqn{ + \epsilon}\out{<sub>x</sub>}, where \eqn{\epsilon}\out{<sub>x</sub>} is the random variable drawn from a normal distribution with mean 0 and SD \eqn{\sigma}\out{<sub>env</sub>}. See \href{https://github.com/aterui/mcbrnet}{github page} for further details.}

\item{sd_env_lon}{numeric value indicating the SD of longitudinal environmental noise (\eqn{\sigma}\out{<sub>env</sub>}).}

\item{randomize_patch}{logical indicating whether randomize patches or not. If \code{FALSE}, the function may generate a biased network with ordered patches. Default \code{TRUE}.}

\item{plot}{logical indicating if a plot should be shown or not. If \code{FALSE}, a plot of the generated network will not be shown. Default \code{TRUE}.}

\item{patch_label}{character string indicating a type of patch (vertex) label (either \code{"patch", "branch", "n_upstream"}). \code{"patch"} shows patch ID, \code{"branch"} branch ID, and \code{"n_upstream"} the number of upstream contributing patches. If \code{NULL}, no label will be shown on patches in the plot. Default \code{NULL}.}

\item{patch_size}{patch (vertex) size in the plot. Default 6.}

\item{patch_scaling}{logical. If \code{TRUE}, patch (vertex) size will be proportional to the number of upstream contributing patches. The patch (vertex) size will be equal to \code{0.3 * scale_factor} at the upstream terminals and \code{1.3 * scale_factor} at the root. Overrides \code{patch_size}.}

\item{scale_factor}{numeric value scaling patch (vertex) size. Enabled if \code{patch_scaling = TRUE}.}
}
\value{
\code{adjacency_matrix} adjacency matrix for the generated network.

\code{distance_matrix} distance matrix for the generated network.

\code{df_patch} a data frame containing patch attributes.
}
\description{
Generate a random branching network
}
\section{Reference}{
 see \href{https://github.com/aterui/mcbrnet}{github page} for instruction
}

\examples{
brnet(n_patch = 100, p_branch = 0.5)

}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
