% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm.R
\name{extra_prey}
\alias{extra_prey}
\title{Extra prey function}
\usage{
extra_prey(alpha0, j, i0, tp, theta, kappa, cannibal = FALSE)
}
\arguments{
\item{alpha0}{Interaction matrix}

\item{j}{Integer. Consumer's index}

\item{i0}{Integer. Index for the first prey}

\item{tp}{Numeric. Initial trophic positions}

\item{theta}{Numeric. Scale parameter of an exponential distribution. Smaller values indicate greater trophic specialization.}

\item{kappa}{Integer. Number of extra prey items}

\item{cannibal}{Logical. If \code{TRUE}, cannibalism allowed}
}
\description{
Extra prey function
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
