% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcbrnet.R
\name{findr}
\alias{findr}
\title{Find suitable intrinsic growth rates}
\usage{
findr(
  alpha,
  k0,
  theta = 1,
  lambda0 = 0.001,
  interval = 0.01,
  sigma = 0.01,
  maxit = 1000
)
}
\arguments{
\item{alpha}{Numeric. n_species x n_species interaction matrix}

\item{k0}{Numeric. Total carrying capacity for basal species combined}

\item{theta}{Numeric. Shape parameter for a Dirichlet distribution that determines relative equilibrium densities of basal species}

\item{lambda0}{Numeric. Initial lambda value for the exponential decay of equilibrium densities with trophic position}

\item{interval}{Numeric. Increment of lambda value.}

\item{sigma}{Numeric. Degree of noise added to equilibrium densities}

\item{maxit}{Integer. Maximum number of iterations to find a suitable lambda}
}
\description{
Find suitable intrinsic growth rates
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
