% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcbrnet.R
\name{frgm}
\alias{frgm}
\title{Dispersal probability matrix after accounting for network fragmentation}
\usage{
frgm(x, rate, pattern = "random", p = NULL, n_barrier)
}
\arguments{
\item{x}{'brnet' object or adjacency matrix}

\item{rate}{Numeric.
Rate parameter of exponential dispersal kernel}

\item{pattern}{Character.
Specify fragmentation pattern ("random", "upstream", or "downstream")}

\item{p}{Numeric.
Passability of fragmented edges (probability, from 0 to 1).}

\item{n_barrier}{Integer.
Number of barriers.}
}
\description{
Dispersal probability matrix after accounting for network fragmentation
}
\section{Reference}{
 see \href{https://aterui.github.io/mcbrnet/}{package webpage} for instruction
}

\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
