% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frgm.R
\name{frgm}
\alias{frgm}
\title{Dispersal probability matrix after accounting for network fragmentation}
\usage{
frgm(x, rate, pattern = "random", p = NULL, n_barrier)
}
\arguments{
\item{x}{'brnet' object or adjacency matrix}

\item{rate}{rate parameter of exponential dispersal kernel}

\item{pattern}{fragmentation pattern}

\item{p}{passability of fragmented edges (probability)}

\item{n_barrier}{number of barriers}
}
\description{
Dispersal probability matrix after accounting for network fragmentation
}
\section{Reference}{
 see \href{https://aterui.github.io/mcbrnet/}{package webpage} for instruction
}

\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
