% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_adj.R
\name{fun_adj}
\alias{fun_adj}
\title{Internal function: return adjacency in a linear habitat}
\usage{
fun_adj(n, start_id = 1)
}
\arguments{
\item{n}{Number of nodes}

\item{start_id}{Start node ID}
}
\description{
Internal function: return adjacency in a linear habitat
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
