% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fun_disp_mat}
\alias{fun_disp_mat}
\title{To dispersal matrix}
\usage{
fun_disp_mat(
  n_patch,
  landscape_size,
  theta,
  xy_coord = NULL,
  distance_matrix = NULL,
  dispersal_matrix = NULL
)
}
\arguments{
\item{n_patch}{Integer.
Number of patches in a metacommunity.}

\item{landscape_size}{Length of a landscape on a side.
Enabled if \code{distance_matrix = NULL} and \code{dispersal_matrix = NULL}.}

\item{theta}{Numeric.
Rate parameter of exponential dispersal kernel.}

\item{xy_coord}{Site coordinates.
Must be provided as a data frame,
in which each row corresponds to an individual site
with x and y coordinates (columns).
Defualt \code{NULL}.}

\item{distance_matrix}{Distance matrix.
Each element must indicate distance between a given pair of habitat patches.
If provided, the distance matrix will be used
to generate dispersal matrix and distance decay of environmental correlations.
Default \code{NULL}.}

\item{dispersal_matrix}{Dispersal matrix.
If provided, this matrix overrides \code{distance_matrix}
to simulate dispersal process.
Default \code{NULL}.}
}
\description{
To dispersal matrix
}
