% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_disp_mat.R
\name{fun_disp_mat}
\alias{fun_disp_mat}
\title{Internal function: to dispersal matrix}
\usage{
fun_disp_mat(
  n_patch,
  landscape_size,
  theta,
  xy_coord = NULL,
  distance_matrix = NULL,
  dispersal_matrix = NULL
)
}
\arguments{
\item{n_patch}{number of patches}

\item{landscape_size}{landscape size}

\item{theta}{rate parameter for a dispersal kernel}

\item{xy_coord}{xy coordinate}

\item{distance_matrix}{distance matrix}

\item{dispersal_matrix}{dispersal matrix}
}
\description{
Internal function: to dispersal matrix
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
