% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_dispersal.R
\name{fun_dispersal}
\alias{fun_dispersal}
\title{Internal function: dispersal}
\usage{
fun_dispersal(x, v_p_dispersal, m_dispersal)
}
\arguments{
\item{x}{Population size matrix}

\item{v_p_dispersal}{Vector of dispersal probability}

\item{m_dispersal}{Dispersal matrix}
}
\description{
Internal function: dispersal
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
