% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get_n_branch.R
\name{fun_get_n_branch}
\alias{fun_get_n_branch}
\title{Internal function: get n_branch}
\usage{
fun_get_n_branch(n_patch, p_branch)
}
\arguments{
\item{n_patch}{number of patches}

\item{p_branch}{branching probability}
}
\description{
Internal function: get n_branch
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
