% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_igp.R
\name{fun_igp}
\alias{fun_igp}
\title{Internal function: igp function}
\usage{
fun_igp(
  x,
  r_b = 5,
  k = 100,
  e = c(5, 5, 5),
  a = c(0.5, 0.5, 0.5),
  h = c(1, 1, 1),
  s0 = rep(0.8, 3)
)
}
\arguments{
\item{x}{Community matrix for basal, intraguild prey (ig-prey), and intraguild predator (ig-predator)}

\item{r_b}{Maximum reproductive rate of basal species}

\item{k}{Carrying capacity}

\item{e}{Energetic conversion efficiency; must be given by the order of basal to ig-prey, basal to ig-predator, and ig-prey to ig-predator}

\item{a}{Attack rate. Must be given by the order of basal to ig_prey, basal to ig-predator, and ig-prey to ig-predator}

\item{h}{Handling time. Must be given by the order of basal to ig_prey, basal to ig-predator, and ig-prey to ig-predator}

\item{s0}{Background survival rate. Must be given by the order of basal, ig-prey, and ig-predator}
}
\description{
Internal function: igp function
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
