% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fun_igp}
\alias{fun_igp}
\title{IGP function}
\usage{
fun_igp(
  x,
  r_b = 5,
  k = 100,
  e = c(0.8, 0.8, 0.8),
  a = c(2, 2, 2),
  h = c(0.1, 0.1, 0.1),
  s = 0
)
}
\arguments{
\item{x}{Community matrix for basal, intraguild prey (ig-prey), and intraguild predator (ig-predator)}

\item{r_b}{Maximum reproductive rate of basal species}

\item{k}{Carrying capacity of basal species}

\item{e}{Energetic conversion efficiency; must be given by the order of basal to ig-prey, basal to ig-predator, and ig-prey to ig-predator}

\item{a}{Attack rate. Must be given by the order of basal to ig_prey, basal to ig-predator, and ig-prey to ig-predator}

\item{h}{Handling time. Must be given by the order of basal to ig_prey, basal to ig-predator, and ig-prey to ig-predator}

\item{s}{Strength of prey switching from ig-prey to basal}
}
\description{
IGP function
}
