% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_int_mat.R
\name{fun_int_mat}
\alias{fun_int_mat}
\title{Internal function: interaction matrix}
\usage{
fun_int_mat(n_species, alpha, min_alpha, max_alpha, interaction_type)
}
\arguments{
\item{n_species}{Number of species}

\item{alpha}{Interaction strength}

\item{min_alpha}{Minimum interaction strength}

\item{max_alpha}{Maximum interaction strength}

\item{interaction_type}{Type of interaction}
}
\description{
Internal function: interaction matrix
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
