% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_m_adj.R
\name{fun_m_adj}
\alias{fun_m_adj}
\title{Internal function: adjacency matrix in branching networks}
\usage{
fun_m_adj(n_patch, p_branch, n_branch, n_patch_free)
}
\arguments{
\item{n_patch}{number of patches}

\item{p_branch}{branching probability}

\item{n_branch}{number of branches}

\item{n_patch_free}{constraint on number of patches}
}
\description{
Internal function: adjacency matrix in branching networks
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
