% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_to_m.R
\name{fun_to_m}
\alias{fun_to_m}
\title{Internal function: to vector}
\usage{
fun_to_m(x, n_species, n_patch, param_attr)
}
\arguments{
\item{x}{scalar or vector}

\item{n_species}{number of species}

\item{n_patch}{number of patches}

\item{param_attr}{indicate species or patch attribute}
}
\description{
Internal function: to vector
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
