% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brnet.R
\name{ggbrnet}
\alias{ggbrnet}
\title{Visualize a branching network}
\usage{
ggbrnet(
  x,
  patch_color = "env",
  edge_color = "black",
  edge_weight = "none",
  value_col = NULL,
  color_label = NULL,
  patch_label = "none",
  patch_size = 3,
  ...
)
}
\arguments{
\item{x}{\code{brnet()} object}

\item{patch_color}{Type of patch (vertex) label (either \code{"env"}, \code{"disturbance"}, \code{"other"} or any color). Default \code{"env"}.}

\item{edge_color}{Edge color}

\item{edge_weight}{Type of edge weight (\code{"passability"}). Default \code{NULL}.}

\item{value_col}{Patch values. Must be specified if \code{patch_color = "other"}.}

\item{color_label}{Color legend title}

\item{patch_label}{Type of patch (vertex) label (either \code{"patch", "branch", "n_upstream"}). \code{"patch"} shows patch ID, \code{"branch"} branch ID, and \code{"n_upstream"} the number of upstream contributing patches. If \code{"none"}, no label will be shown on patches in the plot. Default \code{"none"}.}

\item{patch_size}{Patch (vertex) size in the plot.}

\item{...}{Arguments passed to \code{ggraph::geom_node_label()}.}
}
\description{
Visualize a branching network
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
