% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcbrnet.R
\name{ggbrnet}
\alias{ggbrnet}
\title{Visualize a branching network}
\usage{
ggbrnet(
  x,
  patch_color = "steelblue",
  edge_color = "black",
  edge_alpha = 1,
  edge_linetype = "solid",
  edge_width = 0.5,
  patch_label = "none",
  patch_size = 3,
  ...
)
}
\arguments{
\item{x}{\code{brnet()} object}

\item{patch_color}{Type of patch (vertex) label (either \code{"env"}, \code{"disturb"}, any column name in \code{`df_patch`}, or any color).}

\item{edge_color}{Edge color}

\item{edge_alpha}{Edge transparency. Numeric or one of the columns in \code{df_edge}.}

\item{edge_linetype}{Edge line type}

\item{edge_width}{Edge line width}

\item{patch_label}{Type of patch (vertex) label (either \code{"patch", "branch", "n_upstream"}). \code{"patch"} shows patch ID, \code{"branch"} branch ID, and \code{"n_upstream"} the number of upstream contributing patches. If \code{"none"}, no label will be shown on patches in the plot.}

\item{patch_size}{Patch (vertex) size in the plot.}

\item{...}{Arguments passed to \code{ggraph::geom_node_label()}.}
}
\description{
Visualize a branching network
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
