% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_igpsim.R
\name{igpsim}
\alias{igpsim}
\title{Simulate meta-food web dynamics with intraguild predation}
\usage{
igpsim(
  n_patch = 5,
  n_warmup = 200,
  n_burnin = 200,
  n_timestep = 1000,
  r_b = 5,
  conv_eff = rep(0.5, 3),
  attack_rate = rep(1, 3),
  handling_time = rep(0, 3),
  s = 0,
  propagule_interval = NULL,
  propagule_seed = 0.5,
  carrying_capacity = 100,
  xy_coord = NULL,
  distance_matrix = NULL,
  dispersal_matrix = NULL,
  p_disturb = 0,
  m_disturb = 0,
  landscape_size = 10,
  p_dispersal = 0.1,
  theta = 1,
  plot = FALSE
)
}
\arguments{
\item{n_patch}{Number of patches in a metacommunity.}

\item{n_warmup}{Number of time-steps for warm-up. Default \code{200}.}

\item{n_burnin}{Number of time-steps for burn-in. Default \code{200}.}

\item{n_timestep}{Number of time-steps to be saved. Default \code{1000}.}

\item{r_b}{Maximum reproductive rate of basal species}

\item{conv_eff}{Energetic conversion efficiency; must be given by the order of basal to ig-prey, basal to ig-predator, and ig-prey to ig-predator}

\item{attack_rate}{Attack rate. Must be given by the order of basal to ig_prey, basal to ig-predator, and ig-prey to ig-predator}

\item{handling_time}{Handling time. Must be given by the order of basal to ig_prey, basal to ig-predator, and ig-prey to ig-predator}

\item{s}{Strength of switching between basal and ig-prey}

\item{propagule_interval}{Time interval for propagule introduction during warm-up. If \code{NULL}, a value of \code{ceiling(n_warmup / 10)} will be used.}

\item{propagule_seed}{Propagule mean density (intensity parameter in a Poisson distribution). Default \code{0.5}.}

\item{carrying_capacity}{Carrying capacities of individual patches. Length must be one or equal to \code{n_patch}. Default \code{100}.}

\item{xy_coord}{Data frame for site coordinates. Each row should correspond to an individual patch, with x and y coordinates (columns). Defualt \code{NULL}.}

\item{distance_matrix}{Distance matrix indicating distance between habitat patches. If provided, the distance matrix will be used to generate dispersal matrix and to calculate distance decay of environmental correlations. Default \code{NULL}.}

\item{dispersal_matrix}{Dispersal matrix to be used to simulate dispersal process. Override distance_matrix. Default \code{NULL}.}

\item{p_disturb}{Disturbance probability.}

\item{m_disturb}{Disturbance-induced proportional mortality.}

\item{landscape_size}{Length of a landscape on a side. Enabled if \code{dispersal_matrix = NULL}.}

\item{p_dispersal}{Probability of dispersal. Length must be one or equal to \code{n_species}.}

\item{theta}{Dispersal parameter describing dispersal capability of species.}

\item{plot}{If \code{TRUE}, five sample patches and species of \code{df_dynamics} are plotted.}
}
\description{
Simulate meta-food web dynamics with intraguild predation
}
\section{Reference}{
 see \href{https://aterui.github.io/mcbrnet/}{package webpage} for instruction
}

\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
