% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcbrnet-package.R
\docType{package}
\name{mcbrnet-package}
\alias{mcbrnet}
\alias{mcbrnet-package}
\title{mcbrnet: Simulation functions for metacommunity dynamics in a branching network}
\description{
Simulation functions for metacommunity dynamics in a brancing network.
    Function \code{brnet} generates a random branching network and returns adjacency and distance matrices.
    Function \code{mcsim} simulates metacommunity dynamics. By default, it produces a square-shaped landscape with randomly distributed habitat patches. Matrices generated by \code{brnet} may be used to simulate metacommunity dynamics in a branching network.
    Function \code{adjtodist} converts an adjacency matrix into a distance matrix.
}
\keyword{internal}
