% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcbrnet-package.R
\docType{package}
\name{mcbrnet-package}
\alias{mcbrnet}
\alias{mcbrnet-package}
\title{mcbrnet: Metacommunity simulation in branching networks}
\description{
Metacommunity simulation in branching networks.
}
\keyword{internal}
