% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcbrnet-package.R
\docType{package}
\name{mcbrnet-package}
\alias{mcbrnet}
\alias{mcbrnet-package}
\title{mcbrnet: Simulation functions for metacommunity dynamics in a branching network}
\description{
Simulation functions for metacommunity dynamics in a brancing network.    
    \code{mcsim} Simulate metacommunity dynamics.   
    \code{brnet} Generate a random branching network.
}
\keyword{internal}
