% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_mcsim.R
\name{mcsim}
\alias{mcsim}
\title{Simulate metacommunity dynamics}
\usage{
mcsim(
  n_species = 5,
  n_patch = 5,
  n_warmup = 200,
  n_burnin = 200,
  n_timestep = 1000,
  propagule_interval = NULL,
  carrying_capacity = 100,
  interaction_type = "constant",
  alpha = 0,
  min_alpha = NULL,
  max_alpha = NULL,
  r0 = 4,
  sd_niche_width = 1,
  optim_min = -1,
  optim_max = 1,
  distance_matrix = NULL,
  landscape_size = 10,
  mu_env = 0,
  sd_env = 0.1,
  phi = 1,
  spatial_env_cor = FALSE,
  p_dispersal = 0.1,
  theta = 1,
  plot = FALSE
)
}
\arguments{
\item{n_species}{numeric value. Number of species in a metacommunity.}

\item{n_patch}{numeric value. Number of patches in a metacommunity.}

\item{n_warmup}{numeric value. Number of time-steps for warm-up.}

\item{n_burnin}{numeric value. Number of time-steps for burn-in.}

\item{n_timestep}{numeric value. Number of time-steps to be saved.}

\item{propagule_interval}{numeric value. Time interval for propagule introduction during warm-up. If \code{NULL}, a value of \code{ceiling(n_warmup / 10)} will be used.}

\item{carrying_capacity}{numeric value (length should be equal to one or the number of patches). Carrying capacities of individual patches.}

\item{interaction_type}{character string. \code{"constant"} or \code{"random"}. \code{"constant"} assumes the single interaction strength of alpha for all pairs of species. \code{"random"} draws random numbers from a uniform distribution with \code{min_alpha} and \code{max_alpha}.}

\item{alpha}{numeric value. Species interaction strength.}

\item{min_alpha}{numeric value. Minimum value of a uniform distribution that generates alpha.}

\item{max_alpha}{numeric value. Maximum value of a uniform distribution that generates alpha.}

\item{r0}{numeric value (length should be equal to one or the number of species). Maximum population growth rate of the Beverton-Holt model.}

\item{sd_niche_width}{numeric value. Niche width of species. Higher values indicate greater niche width.}

\item{optim_min}{numeric value. Minimum value of a uniform distribution that generates optimal environmental values of simulated species. Values are randomly assigned to species.}

\item{optim_max}{numeric value. Maximum value of a uniform distribution that generates optimal environmental values of simulated species. Values are randomly assigned to species.}

\item{distance_matrix}{numeric value. Distance matrix indicating distance between habitat patches. If \code{NULL}, a square landscape with randomly distributed patches will be generated. Default \code{NULL}.}

\item{landscape_size}{numeric value. Length of a landscape on a side. Active only when \code{dispersal_matrix = NULL}.}

\item{mu_env}{numeric value (length should be equal to one or the number of patches). Mean environmental values of patches.}

\item{sd_env}{numeric value. Temporal SD of environmental variation at each patch.}

\item{phi}{numeric value. Parameter describing distance decay of spatial autocorrelation in temporal environmental variation.}

\item{spatial_env_cor}{logical. Indicates whether spatial autocorrelation in temporal environmental variation is considered or not. Default \code{FALSE}.}

\item{p_dispersal}{numeric value or vector (length should be equal to one or the number of species). Probability of dispersal.}

\item{theta}{numeric value. Dispersal parameter describing dispersal capability of species.}

\item{plot}{logical. If \code{TRUE}, five sample patches and species of \code{df_dynamics} are plotted.}
}
\value{
\code{df_dynamics} a data frame containing simulated metacommunity dynamics.

\code{df_species} a data frame containing species attributes.

\code{df_patch} a data frame containing patch attributes.

\code{df_diversity} a data frame containing diversity metrics.
}
\description{
Simulate metacommunity dynamics
}
\examples{
mcsim(n_warmup = 200, n_burnin = 200, n_timestep = 1000)

}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
