% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptsource.R
\name{ptsource}
\alias{ptsource}
\title{Create point source disturbance}
\usage{
ptsource(x, n_source, p, q, pattern = "random")
}
\arguments{
\item{x}{\code{brnet()} object}

\item{n_source}{Number of point sources}

\item{p}{Reduction factor for downstream propagation}

\item{q}{Reduction factor for upstream propagation}

\item{pattern}{Spatial pattern of point sources. "random", "cluster", "upstream", "downstream"}
}
\description{
Create point source disturbance
}
\section{Reference}{
 see \href{https://aterui.github.io/mcbrnet/}{package webpage} for instruction
}

\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
