% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sglv.R
\name{sglv}
\alias{sglv}
\title{Spatial generalized Lotka-Volterra model}
\usage{
sglv(
  n_species,
  n_patch,
  n_timestep = 100,
  interval = 0.01,
  r,
  alpha,
  phi,
  connectivity,
  disturb = list(int = 1, rate = (1/n_timestep) * 10, s = interval * 10),
  n0 = list(min = 0, max = 1),
  threshold = 1e-04,
  ...
)
}
\arguments{
\item{n_species}{Integer. Number of species}

\item{n_patch}{Integer. Number of habitat patches}

\item{n_timestep}{Integer. Number of time step for a simulation run}

\item{interval}{Numeric. Time interval used for a numerical ODE solver}

\item{r}{Numeric. Intrinsic growth rates of modeled species}

\item{alpha}{n_species x n_species interaction matrix}

\item{phi}{Numeric. Dispersal rate parameter}

\item{connectivity}{n_patch x n_patch connectivity (adjacency) matrix.}

\item{disturb}{List. Parameters for disturbance: \code{int} = disturbance intensity; \code{rate} = disturbance frequency; \code{s} = temporal duration of disturbance.}

\item{n0}{List. Initial densities for modeled species. Randomly generated by \code{runif(n_species, min = min, max = max)}}

\item{threshold}{Numeric. Extinction threshold. Species will be removed from a simulation if species density goes below this value}

\item{...}{Additional arguments passed to \code{\link{ode}}}
}
\description{
Spatial generalized Lotka-Volterra model
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
