% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm.R
\name{to_alpha}
\alias{to_alpha}
\title{Apply conversion efficiency and attack rate}
\usage{
to_alpha(
  A,
  attack = list(min = 0, max = 1),
  convert = list(min = 0, max = 1),
  mortal = list(min = 0, max = 1)
)
}
\arguments{
\item{A}{Interaction matrix}

\item{attack}{List for attack rates. Specify minimum and maximum values for a uniform distribution.}

\item{convert}{List for conversion efficiency. Specify minimum and maximum values for a uniform distribution.}

\item{mortal}{List for mortality (or intraspecific competition). Specify minimum and maximum values for a uniform distribution.}
}
\description{
Apply conversion efficiency and attack rate
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
